+ Virtualized Platinum/Gold trading for cross-client-version compatibility. + Dropping Bank Checks or Gold in to a trade window will automatically credit your account and increase your offer. * Platinum and Gold amounts are displayed to unsupported clients by the presence of a virtual bank check in the trade window; * This virtual bank check is automatically updated and is only visible to those with unsupported clients.

+++ Scripts/Accounting/Account.cs
@@ -1285,7 +1285,7 @@ namespace Server.Accounting
 		///     By default, when 1,000,000,000 Gold is accumulated, it will transform
 		///     into 1 Platinum.
 		/// </summary>
-		public static double CurrencyThreshold = 1000000000;
+		public static int CurrencyThreshold = 1000000000;
 
 		/// <summary>
 		///     This amount represents the total amount of currency owned by the player.
